book_store/gateway/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pyjwt
redis
fakeredis